feat(plugin-debug): expose introspect-mcp ToolsExplorer as a devtools panel
- Factored the `ToolsExplorer` component out of
  `react-ui-introspect/Tools.stories.tsx` into
  `components/ToolsExplorer/ToolsExplorer.tsx` (and renamed the stories
  file alongside). The component is now a real export of
  `@dxos/react-ui-introspect`; the stories file is a thin wrapper that
  just sets the storybook meta. `DEFAULT_INTROSPECT_MCP_URL` constant
  exposes the configured server URL.
- Added a new devtools graph node + Surface in `plugin-debug`,
  matching the existing `app-graph` pattern: `Devtools.ToolsExplorer`
  literal, `'tools-explorer'` graph node under the devtools root with
  the `ph--toolbox--regular` icon and `'debug-tools-explorer.label'`
  translation, and an `AppSurface.literal` filter that renders
  `<ToolsExplorer />` for the Article role.
- Added `@modelcontextprotocol/sdk` and `@dxos/introspect-tools` to
  `react-ui-introspect`'s runtime deps (they were dev-only when only
  reachable from a story); added `@dxos/react-ui-introspect` to
  `plugin-debug`'s deps.
