+++ packages/plugins/plugin-debug/src/translations.ts
@@ -19,6 +19,7 @@ export const translations = [
         'debug.label': 'Debug',
         'generate-objects.label': 'Generate Objects',
         'debug-app-graph.label': 'App Graph',
+        'debug-tools-explorer.label': 'Tools Explorer',
 
         'settings.wireframe.label': 'Show wireframes.',
         'settings.wireframe.description': 'Overlay wireframe outlines on UI components for debugging layout.',
