+++ packages/plugins/plugin-debug/tsconfig.json
@@ -120,6 +120,9 @@
       "path": "../../ui/react-ui-form"
     },
     {
+      "path": "../../ui/react-ui-introspect"
+    },
+    {
       "path": "../../ui/react-ui-mosaic"
     },
     {
