+++ packages/ui/react-ui-introspect/src/components/ToolsExplorer/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+export * from './ToolsExplorer';
