+++ packages/ui/react-ui-introspect/src/components/index.ts
@@ -5,4 +5,5 @@
 export * from './ToolList';
 export * from './ToolForm';
 export * from './ToolResults';
+export * from './ToolsExplorer';
 export type { ToolEntry } from './types';
